Resolvendo problemas com biblioteca Ymodem
